Force abapcleaner.gui plugin activation when checking handler enablement
Eclipse is optimistic with regards to enablement of a Handler, i.e. it
will not activate a plugin to call AbstractHandler#setEnabled().

This can lead to situations where the ABAP cleaner actions in the menus
are enabled, even if they are not applicable to the object type in the
editor, because the com.sap.adt.abapcleaner.gui plugin is not yet fully
loaded.

This change adds a dummy property tester to enforce the plugin
activation.